renamed test_cases into scenarii
