+++ src/pyosrd/scenarii/c2y1y2y_2_trains_no_conflict.py
@@ -1,7 +1,7 @@
-from rlway.pyosrd import OSRD
+from pyosrd import OSRD
 
 
-def c2y1y2y_2_trains_same_order() -> OSRD:
+def c2y1y2y_2_trains_no_conflict() -> OSRD:
     """
     See c2y1y2y.py in use_cases to check the infra. (reminder here :)
        station0 (2 tracks)                        station1 (2 tracks)                      station2 (1 track)
@@ -16,14 +16,14 @@ def c2y1y2y_2_trains_same_order() -> OSRD:
     train0 goes from T0 to T7 (passing by T4)
     train1 foes from T1 to T7 (passing by T5)
 
-    In this test case train0 is slightly delayed BUT train1 should be stopped a little bit (but not overtake)
+    In this test case train0 is slightly delayed which should not delay train1
 
     """  # noqa
     use_case = 'c2y1y2y'
     sim = OSRD(use_case=use_case, dir='tmp')
 
     sim.reset_delays()
-    sim.add_delay('train0', time_threshold=130, delay=75.)
+    sim.add_delay('train0', time_threshold=130, delay=10.)
     sim.run()
 
     return sim
